refactor: make agent root a facade
